add remaining form info
