+++ app/components/add-feed-tabs/add-feed-tab/component.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 
 export default Ember.Component.extend({
-	tagName: 'button',
+	// tagName: 'button',
 	classNames: 'btn',
 });
