+++ app/components/add-feed-tabs/add-feed-tab/template.hbs
@@ -1,3 +1,2 @@
-<div class="add-feed-nav-tabs">
-	{{tab-text}}
-</div>
+
+<div class={{tab-class}}>{{tab-text}}</div>
