+++ app/feeds/new/index/template.hbs
@@ -1,0 +1,28 @@
+<div class="fullvw-wrapper">
+	{{add-feed-tabs header-text="ADD A TRANSIT FEED"}}
+</div>
+<div class="container">
+	<p>Need blurb here for upload a feed.</p>
+</div>
+<div class="container email-input-form">
+	<div class="row">
+		<div class="col-sm-5 input-section">
+			<div class="input-form-lable">Pase a .GTFS feed URL</div>
+			<div class="input-form-field" {{input type="text" class="form-control" value=xxx}}>http://...</div>
+		</div>
+	</div>	
+</div>
+<br>
+<div class="container email-caption">
+	Transitland will periodically check this URL for updates
+</div>
+<br>
+<div align="left">
+	{{#link-to "index" class="btn btn-default" role="button"}}
+			Cancel
+	{{/link-to}}
+
+	{{#link-to "feeds.new.add-operator" class="btn btn-default" role="button"}}
+			Next >
+	{{/link-to}}
+</div>
