+++ app/feeds/new/license/template.hbs
@@ -1,0 +1,44 @@
+<div class="fullvw-wrapper">
+	{{add-feed-tabs header-text="ADD A TRANSIT FEED" submitted=false}}
+</div>
+
+<div>
+	{{add-feed-blurb blurb-text="Need blurb for add license form instructions."}}
+</div>
+<div class="container input-form">
+	<div class="row">
+		<div class="col-sm-7 col-xs-12 input-section">
+			<div class="input-form-lable">License name</div>
+			<div class="input-form-field" {{input type="text" class="form-control" value=operator.name}}>e.g. Creative Commons 2.0</div>
+		</div>
+		<div class="col-sm-5 col-xs-12 input-section">
+			<div class="input-form-lable">License URL</div>
+			<div class="input-form-field" {{input type="text" class="form-control" value=operator.short_name}}>http://...</div>
+		</div>
+	</div>
+	<div class="row">
+		<div class="col-sm-12 col-xs-12 input-form-guide input-section">Pick the license restrictions. Select 'Unknown'if you are not sure.
+		</div>
+	</div>
+	<div class="row">
+		<div class="col-sm-12 col-xs-12">Attribution for feed data
+			<group class="inline-radio">
+				<div><input type="radio" name="attribution"><label>Optional</label></div>
+				<div><input type="radio" name="attribution"><label>Required</label></div>
+				<div><input type="radio" name="attribution" checked><label>Unknown</label></div>
+			</group>
+		</div>
+	</div>	
+</div>
+
+<br>
+<div align=left>
+{{#link-to "feeds.new.index" class="btn btn-default" role="button"}}
+  Back
+{{/link-to}}
+</div>
+<div align=right>
+{{#link-to "feeds.new.submit" class="btn btn-default" role="button"}}
+	SUBMIT FEED
+{{/link-to}}
+</div>
