--- app/feeds/template.hbs
@@ -1,1 +1,0 @@
-{{outlet}}
