+++ app/index/template.hbs
@@ -12,7 +12,7 @@
 
 
 {{#if editingMode}} 
-	{{#link-to "operators.new" class="btn btn-default" role="button"}}
-		Add an Operator
+	{{#link-to "feeds.new" class="btn btn-default" role="button"}}
+		Add a Feed
 	{{/link-to}}
 {{/if}}
