+++ app/styles/app.scss
@@ -5,4 +5,5 @@
         "shared/footer",
         "operator",
         "operators",
+        "feeds",
         "pod-styles";
